+++ src/pharmpy/modeling/parameters.py
@@ -1,7 +1,8 @@
 from collections.abc import Mapping
 from typing import Optional, Union
 
-from pharmpy.model import Model, Parameter, Parameters, RandomVariables
+from pharmpy.basic import Expr
+from pharmpy.model import Assignment, Model, Parameter, Parameters, RandomVariables
 
 
 def get_thetas(model: Model):
@@ -673,14 +674,17 @@ def replace_fixed_thetas(model: Model):
     """
 
     keep = []
-    d = {}
+    new_assignments = []
 
     for p in model.parameters:
         if p.fix:
-            d[p.symbol] = p.init
+            ass = Assignment(p.symbol, Expr.float(p.init))
+            new_assignments.append(ass)
         else:
             keep.append(p)
 
-    model = model.replace(parameters=Parameters(keep), statements=model.statements.subs(d))
+    model = model.replace(
+        parameters=Parameters(keep), statements=new_assignments + model.statements
+    )
     model = model.update_source()
     return model